[Refactor] authwrapper 수정
