+++ app/_components/commons/common-sections/user-info/UserInfoSection.tsx
@@ -4,8 +4,11 @@ import RatingStar from "@/public/images/icons/icon_star.svg";
 import styles from "./UserInfoSection.module.css";
 import { UserProfileResponseDto } from "@/backend/users/user/applications/dtos/UserDtos";
 
+// 프론트엔드에서 id가 string일 수도 있으므로 확장 타입 정의
+type UserInfoSectionData = Omit<UserProfileResponseDto, 'id'> & { id: string | number };
+
 interface UserInfoSectionProps {
-  data: UserProfileResponseDto;
+  data: UserInfoSectionData;
   userRole: string;
 }
